Rename digiarch.py -> cli.py, fix code smells found by prospector
